fix: refractor stories from feed as a separate service
