+++ lib/features/feed/presentation/screens/home_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:instagram/features/feed/presentations/screens/feed_screen.dart';
+import 'package:instagram/features/feed/presentation/screens/feed_screen.dart';
 import 'package:instagram/features/reels/presentations/screens/reel_screen.dart';
 import 'package:instagram/features/user/presentations/screens/user_profile.dart';
 import '../../../../core/constants/home_nav_constants.dart';
